// prismicio-types.d.ts
type Simplify<T> = { [KeyType in keyof T]: T[KeyType] };
// ...
interface ProjectPostDocumentData {
  /**
   * Client Name field in *Project*
   *
   * - **Field Type**: Text
   * - **Placeholder**: Please Enter Client's Name (Must)
   * - **API ID Path**: project_post.client_name
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/field#key-text
   */
  client_name: prismic.KeyTextField;

  /**
   * Description field in *Project*
   *
   * - **Field Type**: Rich Text
   * - **Placeholder**: Enter Description (Must)
   * - **API ID Path**: project_post.description
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/field#rich-text-title
   */
  description: prismic.RichTextField;

  /**
   * Project Main Image field in *Project*
   *
   * - **Field Type**: Image
   * - **Placeholder**: *None*
   * - **API ID Path**: project_post.project_main_image
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/field#image
   */
  project_main_image: prismic.ImageField<never>;

  /**
   * Year field in *Project*
   *
   * - **Field Type**: Text
   * - **Placeholder**: Please Enter Year (Must)
   * - **API ID Path**: project_post.year
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/field#key-text
   */
  year: prismic.KeyTextField;

  /**
   * Sector field in *Project*
   *
   * - **Field Type**: Content Relationship
   * - **Placeholder**: *None*
   * - **API ID Path**: project_post.sector
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/field#link-content-relationship
   */
  sector: prismic.ContentRelationshipField<"sector_post">;

  /**
   * Subsector field in *Project*
   *
   * - **Field Type**: Content Relationship
   * - **Placeholder**: *None*
   * - **API ID Path**: project_post.subsector
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/field#link-content-relationship
   */
  subsector: prismic.ContentRelationshipField<"subsector_post">;

  /**
   * Role field in *Project*
   *
   * - **Field Type**: Text
   * - **Placeholder**: Please Enter Role (Must)
   * - **API ID Path**: project_post.role
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/field#key-text
   */
  role: prismic.KeyTextField;

  /**
   * Location field in *Project*
   *
   * - **Field Type**: Text
   * - **Placeholder**: Please Enter Location (Must)
   * - **API ID Path**: project_post.location
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/field#key-text
   */
  location: prismic.KeyTextField;

  /**
   * Project's Gallery field in *Project*
   *
   * - **Field Type**: Group
   * - **Placeholder**: *None*
   * - **API ID Path**: project_post.gallery_images[]
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/field#group
   */
  gallery_images: prismic.GroupField<
    Simplify<ProjectPostDocumentDataGalleryImagesItem>
  >;

  /**
   * Next Project field in *Project*
   *
   * - **Field Type**: Content Relationship
   * - **Placeholder**: *None*
   * - **API ID Path**: project_post.next_project
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/field#link-content-relationship
   */
  next_project: prismic.ContentRelationshipField<"project_post"> /**
   * Meta Title field in *Project*
   *
   * - **Field Type**: Text
   * - **Placeholder**: A title of the page used for social media and search engines
   * - **API ID Path**: project_post.meta_title
   * - **Tab**: SEO & Metadata
   * - **Documentation**: https://prismic.io/docs/field#key-text
   */;
  meta_title: prismic.KeyTextField;

  /**
   * Meta Description field in *Project*
   *
   * - **Field Type**: Text
   * - **Placeholder**: A brief summary of the page
   * - **API ID Path**: project_post.meta_description
   * - **Tab**: SEO & Metadata
   * - **Documentation**: https://prismic.io/docs/field#key-text
   */
  meta_description: prismic.KeyTextField;

  /**
   * Meta Image field in *Project*
   *
   * - **Field Type**: Image
   * - **Placeholder**: *None*
   * - **API ID Path**: project_post.meta_image
   * - **Tab**: SEO & Metadata
   * - **Documentation**: https://prismic.io/docs/field#image
   */
  meta_image: prismic.ImageField<never>;
}

/**
 * Project document from Prismic
 *
 * - **API ID**: `project_post`
 * - **Repeatable**: `true`
 * - **Documentation**: https://prismic.io/docs/custom-types
 *
 * @typeParam Lang - Language API ID of the document.
 */
export type ProjectPostDocument<Lang extends string = string> =
  prismic.PrismicDocumentWithUID<
    Simplify<ProjectPostDocumentData>,
    "project_post",
    Lang
  >;
// ...
export interface SettingsDocumentDataMainNavigationItem {
  /**
   * Navigation Title field in *Settings → Main Navigation*
   *
   * - **Field Type**: Text
   * - **Placeholder**: e.g. Products, Who We Are
   * - **API ID Path**: settings.main_navigation[].title
   * - **Documentation**: https://prismic.io/docs/field#key-text
   */
  title: prismic.KeyTextField;

  /**
   * Navigation Link field in *Settings → Main Navigation*
   *
   * - **Field Type**: Content Relationship
   * - **Placeholder**: *None*
   * - **API ID Path**: settings.main_navigation[].link
   * - **Documentation**: https://prismic.io/docs/field#link-content-relationship
   */
  link: prismic.ContentRelationshipField<
    "page" | "sector_post" | "project_post"
  >;

  /**
   * Has Submenu field in *Settings → Main Navigation*
   *
   * - **Field Type**: Boolean
   * - **Placeholder**: *None*
   * - **Default Value**: false
   * - **API ID Path**: settings.main_navigation[].has_submenu
   * - **Documentation**: https://prismic.io/docs/field#boolean
   */
  has_submenu: prismic.BooleanField;
}
// ...
export interface SettingsDocumentDataSubmenuSectionsItem {
  /**
   * Parent Navigation field in *Settings → Submenu Sections*
   *
   * - **Field Type**: Text
   * - **Placeholder**: Enter exact navigation title (e.g. Who We Are)
   * - **API ID Path**: settings.submenu_sections[].parent_nav
   * - **Documentation**: https://prismic.io/docs/field#key-text
   */
  parent_nav: prismic.KeyTextField;

  /**
   * Submenu Items field in *Settings → Submenu Sections*
   *
   * - **Field Type**: Group
   * - **Placeholder**: *None*
   * - **API ID Path**: settings.submenu_sections[].items[]
   * - **Documentation**: https://prismic.io/docs/field#group
   */
  items: prismic.NestedGroupField<
    Simplify<SettingsDocumentDataSubmenuSectionsItemsItem>
  >;
}

/**
 * Item in *Settings → Featured Sectors in Navigation*
 */
export interface SettingsDocumentDataFeaturedSectorsItem {
  /**
   * Sector field in *Settings → Featured Sectors in Navigation*
   *
   * - **Field Type**: Content Relationship
   * - **Placeholder**: *None*
   * - **API ID Path**: settings.featured_sectors[].sector
   * - **Documentation**: https://prismic.io/docs/field#link-content-relationship
   */
  sector: prismic.ContentRelationshipField<"sector_post">;

  /**
   * Display Subsectors field in *Settings → Featured Sectors in Navigation*
   *
   * - **Field Type**: Boolean
   * - **Placeholder**: *None*
   * - **Default Value**: true
   * - **API ID Path**: settings.featured_sectors[].display_subsectors
   * - **Documentation**: https://prismic.io/docs/field#boolean
   */
  display_subsectors: prismic.BooleanField;

  /**
   * Display Order field in *Settings → Featured Sectors in Navigation*
   *
   * - **Field Type**: Number
   * - **Placeholder**: 1, 2, 3...
   * - **API ID Path**: settings.featured_sectors[].order
   * - **Documentation**: https://prismic.io/docs/field#number
   */
  order: prismic.NumberField;
}

/**
 * Item in *Settings → Product Categories*
 */
export interface SettingsDocumentDataProductCategoriesItem {
  /**
   * Project field in *Settings → Product Categories*
   *
   * - **Field Type**: Content Relationship
   * - **Placeholder**: *None*
   * - **API ID Path**: settings.product_categories[].project
   * - **Documentation**: https://prismic.io/docs/field#link-content-relationship
   */
  project: prismic.ContentRelationshipField<"project_post">;
}
// ...
interface SettingsDocumentData {
  /**
   * Site Title field in *Settings*
   *
   * - **Field Type**: Text
   * - **Placeholder**: Enter site title
   * - **API ID Path**: settings.site_title
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/field#key-text
   */
  site_title: prismic.KeyTextField;

  /**
   * Logo field in *Settings*
   *
   * - **Field Type**: Image
   * - **Placeholder**: *None*
   * - **API ID Path**: settings.site_logo
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/field#image
   */
  site_logo: prismic.ImageField<never>;

  /**
   * Main Navigation field in *Settings*
   *
   * - **Field Type**: Group
   * - **Placeholder**: *None*
   * - **API ID Path**: settings.main_navigation[]
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/field#group
   */
  main_navigation: prismic.GroupField<
    Simplify<SettingsDocumentDataMainNavigationItem>
  >;

  /**
   * Submenu Sections field in *Settings*
   *
   * - **Field Type**: Group
   * - **Placeholder**: *None*
   * - **API ID Path**: settings.submenu_sections[]
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/field#group
   */
  submenu_sections: prismic.GroupField<
    Simplify<SettingsDocumentDataSubmenuSectionsItem>
  >;

  /**
   * Featured Sectors in Navigation field in *Settings*
   *
   * - **Field Type**: Group
   * - **Placeholder**: *None*
   * - **API ID Path**: settings.featured_sectors[]
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/field#group
   */
  featured_sectors: prismic.GroupField<
    Simplify<SettingsDocumentDataFeaturedSectorsItem>
  >;

  /**
   * Product Categories field in *Settings*
   *
   * - **Field Type**: Group
   * - **Placeholder**: *None*
   * - **API ID Path**: settings.product_categories[]
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/field#group
   */
  product_categories: prismic.GroupField<
    Simplify<SettingsDocumentDataProductCategoriesItem>
  >;

  /**
   * Social Media Links field in *Settings*
   *
   * - **Field Type**: Group
   * - **Placeholder**: *None*
   * - **API ID Path**: settings.social_links[]
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/field#group
   */
  social_links: prismic.GroupField<
    Simplify<SettingsDocumentDataSocialLinksItem>
  >;

  /**
   * Contact Information field in *Settings*
   *
   * - **Field Type**: Group
   * - **Placeholder**: *None*
   * - **API ID Path**: settings.contact_information[]
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/field#group
   */
  contact_information: prismic.GroupField<
    Simplify<SettingsDocumentDataContactInformationItem>
  >;

  /**
   * Footer Tagline field in *Settings*
   *
   * - **Field Type**: Text
   * - **Placeholder**: MEP Solutions // Firefighting Solutions // Valves // Pipes // Hose Cabinet // Pumps
   * - **API ID Path**: settings.footer_tagline
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/field#key-text
   */
  footer_tagline: prismic.KeyTextField;

  /**
   * Footer Logo field in *Settings*
   *
   * - **Field Type**: Image
   * - **Placeholder**: *None*
   * - **API ID Path**: settings.footer_logo
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/field#image
   */
  footer_logo: prismic.ImageField<never>;

  /**
   * Footer Legal Links field in *Settings*
   *
   * - **Field Type**: Group
   * - **Placeholder**: *None*
   * - **API ID Path**: settings.footer_legal_links[]
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/field#group
   */
  footer_legal_links: prismic.GroupField<
    Simplify<SettingsDocumentDataFooterLegalLinksItem>
  >;
}

/**
 * Settings document from Prismic
 *
 * - **API ID**: `settings`
 * - **Repeatable**: `false`
 * - **Documentation**: https://prismic.io/docs/custom-types
 *
 * @typeParam Lang - Language API ID of the document.
 */
export type SettingsDocument<Lang extends string = string> =
  prismic.PrismicDocumentWithoutUID<
    Simplify<SettingsDocumentData>,
    "settings",
    Lang
  >;